+++ src/utils.rs
@@ -1,0 +1,5 @@
+use rand::random;
+
+pub fn get_random_positive() -> f64 {
+    random()
+}
